+++ CHANGELOG.md
@@ -26,6 +26,9 @@ We haven't had the best track record of code/feature changes before this date, b
     <summary>
      Unreleased changes that have landed in master. Click to see more.
     </summary>
+
+-  Fix version in header commments
+
   </details>
 
 - 0.10.6: